[CL-2088] Delete module
